Inline the useEthersSigner implementation and delete the hook
